Update slugify logic to handle non-latin characters
